Refactored to characterlist component
